Modify cuda memory threshold

+++ oneflow/core/framework/config_dtr.h
@@ -25,7 +25,7 @@ limitations under the License.
 
 namespace oneflow {
 
-Maybe<void> EnableDTRStrategy(bool enable_dtr, double thres, bool enable_debug);
+Maybe<void> EnableDTRStrategy(bool enable_dtr, size_t thres, bool enable_debug);
 
 }  // namespace oneflow
