+++ oneflow/core/job/job_build_and_infer_ctx_mgr.cpp
@@ -105,7 +105,7 @@ Maybe<void> EagerJobBuildAndInferCtxMgr::VirtualCloseJob() {
 
 bool EagerExecutionEnabled() { return *Global<bool, EagerExecution>::Get(); }
 bool DTREnabled() { return *Global<bool, EnableDTR>::Get(); }
-double GetDTRMemoryThreshold() { return *Global<double, DTRMemoryThreshold>::Get(); }
+size_t GetDTRMemoryThreshold() { return *Global<size_t, DTRMemoryThreshold>::Get(); }
 size_t GetDTRRemainMemory() { return *Global<size_t, DTRRemainMemory>::Get(); }
 bool DTRDebugEnabled() { return *Global<bool, EnableDTRDebug>::Get(); }
